Move useChat to provider
