+++ apps/web/src/features/deep-agent-chat/providers/ChatProvider.tsx
@@ -1,33 +1,16 @@
 "use client";
 
-import React, { createContext, useContext, ReactNode } from "react";
-import { Message, Assistant, Checkpoint, Interrupt } from "@langchain/langgraph-sdk";
+import React, { ReactNode } from "react";
+import { Assistant } from "@langchain/langgraph-sdk";
 import { useChat } from "../hooks/useChat";
 import type { TodoItem } from "../types";
-
-interface ChatContextType {
-  messages: Message[];
-  isLoading: boolean;
-  interrupt: Interrupt | undefined;
-  getMessagesMetadata: (message: Message, index?: number) => any;
-  sendMessage: (message: string) => void;
-  runSingleStep: (
-    messages: Message[],
-    checkpoint?: Checkpoint,
-    isRerunningSubagent?: boolean,
-    optimisticMessages?: Message[]
-  ) => void;
-  continueStream: (hasTaskToolCall?: boolean) => void;
-  stopStream: () => void;
-}
-
-const ChatContext = createContext<ChatContextType | undefined>(undefined);
+import { ChatContext } from "./ChatContext";
 
 interface ChatProviderProps {
   children: ReactNode;
   threadId: string | null;
   setThreadId: (
-    value: string | ((old: string | null) => string | null) | null
+    value: string | ((old: string | null) => string | null) | null,
   ) => void;
   setTodos: (todos: TodoItem[]) => void;
   setFiles: (files: Record<string, string>) => void;
@@ -53,16 +36,8 @@ export function ChatProvider({
     setFiles,
     activeAssistant,
     deploymentId,
-    agentId
+    agentId,
   );
 
   return <ChatContext.Provider value={chat}>{children}</ChatContext.Provider>;
-}
-
-export function useChatContext() {
-  const context = useContext(ChatContext);
-  if (context === undefined) {
-    throw new Error("useChatContext must be used within a ChatProvider");
-  }
-  return context;
 }
